requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2022.4
sqlparse==0.4.3
